Move computeRouteTail to common URL module
